fix: update test expectations for improved 'up' strategy and fix shopping_list type
- Update test_find_best_package_up_strategy: expect minimal overage selection
  - 250g with [100, 250] should pick 250g (0g overage), not 100g (50g overage)
- Fix weekly_plan.py: change shopping_list type from Dict[str, int] to Dict[str, float]
  - Ingredient amounts are floats, matching recipe_db.py and shoplist calculations
  - Update both TypedDict definition and variable annotation

+++ core/weekly_plan.py
@@ -83,7 +83,7 @@ def generate_weekly_plan(
         weekly_coverage[micro] = sum(coverages) / len(coverages)
 
     # Generate shopping list (simple implementation)
-    shopping_list: Dict[str, int] = {}
+    shopping_list: Dict[str, float] = {}
     for day in days:
         for meal in day["meals"]:
             if "ingredients" in meal:
